Ensure security of production deployment by requiring SECRET_KEY.
Reverts a previous change that provided a hard-coded default value for
SECRET_KEY. The dev environment still works fine after this change, since
docker-compose.yml already has a value set for this variable. This only
left the collectstatic call in the Dockerfile, where we just set a value
explicitly for that command.