fix: generalizes player search query format
Prior to this commit, the search query was used monolithically as a
single string to match the members' member id or their (comma-separated)
full name;  this had many shortcomings, e.g., a search query " 12345 "
would fail to match a member with id "12345", and a search query
"surname name" would fail to match a member with full name "surname,
name".  Also see #164 and #248.

In this commit, the search is first stripped of outer whitespace, which
helps search for member ids.  If the search query is not an id, then the
search query is split into tokens, and a match with a member is found if
all of the respective query tokens match the member's full name.  This
allows users to run queries such as "name", "surname", "name surname" or
"surname name".